Correction from git pull request
